 requirements.txt
# requirements.txt
requests
pandas
google-cloud-storage
openpyxl
certifi